+++ app/Livewire/Permisos.php
@@ -6,6 +6,7 @@ use Livewire\Component;
 use Livewire\Attributes\On;
 use Spatie\Permission\Models\Permission;
 use Livewire\WithPagination;
+
 class Permisos extends Component
 {
 
@@ -35,19 +36,20 @@ class Permisos extends Component
         $this->pageTitle = 'Listado';
         $this->componentName = 'Permisos';
     }
+
     public function render()
     {
         $query = Permission::orderBy('id', 'desc');
-        
-        if(strlen($this->search) > 0) {
+
+        if (strlen($this->search) > 0) {
             $query->where('name', 'like', '%' . $this->search . '%');
         }
 
         $data = $query->paginate($this->pagination);
 
-        return view('livewire.permisos.components',['permisos' => $data])
-        ->extends('layouts.app')
-        ->section('content'); 
+        return view('livewire.permisos.components', ['permisos' => $data])
+            ->extends('layouts.app')
+            ->section('content');
     }
 
     public function openModal()
@@ -67,14 +69,20 @@ class Permisos extends Component
 
     public function store()
     {
-        // Validación de reglas
-        $this->validate();
+        try {
+            // Validación de reglas
+            $this->validate();
+            $this->authorize('create', Permission::class);
 
-        Permission::create([
-            'name' => $this->permissionName
-        ]);
+            Permission::create([
+                'name' => $this->permissionName
+            ]);
 
-        $this->dispatch('noty-added', type: 'PERMISO', name: $this->permissionName);
+            $this->dispatch('noty-added', type: 'PERMISO', name: $this->permissionName);
+        } catch (\Illuminate\Auth\Access\AuthorizationException $exception) {
+            // Notificación de error de autorización
+            $this->dispatch('noty-permission', type: 'USUARIO', name: 'PERMISOS', permission: 'CREAR');
+        }
     }
 
     public function edit(Permission $permiso)
@@ -88,36 +96,46 @@ class Permisos extends Component
 
     public function update()
     {
-        // Actualización de reglas de validación para la edición
-        $this->rules['permissionName'] = "required|min:3|unique:permissions,name, {$this->selected_id}";
+        try {
+            // Actualización de reglas de validación para la edición
+            $this->rules['permissionName'] = "required|min:3|unique:permissions,name, {$this->selected_id}";
 
-        // Validación
-        $this->validate();
+            // Validación
+            $this->validate();
+            $this->authorize('update', $this->selected_id);
 
-        $permiso = Permission::find($this->selected_id);
-        $permiso->name = $this->permissionName;
-        $permiso->save();
+            $permiso = Permission::find($this->selected_id);
+            $permiso->name = $this->permissionName;
+            $permiso->save();
 
-        $this->dispatch('noty-updated', type: 'PERMISO', name: $permiso->name );
+            $this->dispatch('noty-updated', type: 'PERMISO', name: $permiso->name);
+        } catch (\Illuminate\Auth\Access\AuthorizationException $exception) {
+            // Notificación de error de autorización
+            $this->dispatch('noty-permission', type: 'USUARIO', name: 'PERMISOS', permission: 'ACTUALIZAR');
+        }
     }
 
     public function destroy($id)
     {
-        $rolesCount = Permission::find($id)->getRoleNames()->count();
-        if($rolesCount > 0)
-        {
-            $this->dispatch('showNotification', 'No se puede eliminar el role porque tiene permisos asociados', 'warning');
-            return;
-        }
+        try {
+            $this->authorize('delete', $id);
+            $rolesCount = Permission::find($id)->getRoleNames()->count();
+            if ($rolesCount > 0) {
+                $this->dispatch('showNotification', 'No se puede eliminar el role porque tiene permisos asociados', 'warning');
+                return;
+            }
 
-        $permiso = Permission::find($id);
-        Permission::find($id)->delete();
-        $this->dispatch('noty-deleted', type: 'PERMISO', name:  $permiso->name);
+            $permiso = Permission::find($id);
+            Permission::find($id)->delete();
+            $this->dispatch('noty-deleted', type: 'PERMISO', name: $permiso->name);
+        } catch (\Illuminate\Auth\Access\AuthorizationException $exception) {
+            $this->dispatch('noty-permission', type: 'USUARIO', name: 'PERMISOS', permission: 'ELIMINAR');
+        }
     }
 
     protected $listeners = [
         'deleteRow' => 'destroy',
-        'searchUpdated' => 'updateSearch', 
+        'searchUpdated' => 'updateSearch',
     ];
 
     public function resetUI()
@@ -127,6 +145,7 @@ class Permisos extends Component
         $this->selected_id = 0;
         $this->resetValidation();
     }
+
     public function updateSearch($search)
     {
         $this->search = $search;
